requirements.txt
boto3==1.4.1
botocore==1.4.61
click==6.6
docutils==0.12
futures==3.0.5
jmespath==0.9.0
pyaml==15.8.2
python-dateutil==2.5.3
PyYAML==3.11
six==1.10.0
pytest==3.0.7
